[#93305592 #93345772] * fixed condition collision when server respond with non keep alive connection header and malformed data at once * documentation clean up according to clang warnings * switch from dispatch_data_t to NSMutableData as read buffer

+++ PubNub/PubNub/PubNub/Misc/Categories/NSString+PNAddition.m
@@ -144,7 +144,7 @@
                     NSUInteger maximumHalfLength = (NSUInteger)ceilf(length * 0.5f);
                     if (maximumHalfLength == (self.length * 0.5f)) {
 
-                        maximumHalfLength = MAX(maximumHalfLength - 2, MAX(maximumHalfLength - 1, 0));
+                        maximumHalfLength = MAX(maximumHalfLength - 2, MAX(maximumHalfLength - 1, (NSUInteger)0));
                     }
 
                     truncatedString = [[NSString alloc] initWithFormat:@"%@…%@", [self substringToIndex:maximumHalfLength],
